requirements.txt
# LLM APIs
anthropic>=0.18.0
openai>=1.12.0

# Vector database
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.3.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development
pytest>=8.0.0